+++ src/data/projects/constants.ts
@@ -13,6 +13,8 @@ export const GH_AUDIO_FINGERPRINTER =
   "https://github.com/BehindTheMusicTree/audio-fingerprinter";
 export const GH_HTMT =
   "https://github.com/BehindTheMusicTree/hear-the-music-tree-api";
+export const GH_PIPELINES =
+  "https://github.com/BehindTheMusicTree/the-music-tree-pipelines";
 
 export function shield(path: string): string {
   return `https://img.shields.io/${path}?style=flat-square`;